separate KE and TKE
