+++ test/runtests.jl
@@ -11,8 +11,12 @@ group     = get(ENV, "TEST_GROUP", :all) |> Symbol
         include("test_tracer_diagnostics.jl")
     end
 
+    if group == :ke_diagnostics || group == :all
+        include("test_kinetic_energy_equation.jl")
+    end
+
     if group == :tke_diagnostics || group == :all
-        include("test_tke_diagnostics.jl")
+        include("test_turbulent_kinetic_energy_equation.jl")
     end
 
     if group == :pe_diagnostics || group == :all
